CHANGE run performance tests without dev-mode
