+++ package.json
@@ -79,9 +79,16 @@
     "test:typings:ci": "npm run transpile && mocha --config ./config/.mocharc.js ./test_tmp/typings.test.js",
     "test:deps": "npm run build && dependency-check ./package.json ./dist/lib/index.js ./dist/lib/plugins/validate-is-my-json-valid.js ./dist/lib/plugins/validate-ajv.js ./dist/lib/plugins/update.js ./dist/lib/plugins/key-compression.js ./dist/lib/plugins/dev-mode/index.js ./dist/lib/plugins/encryption.js ./dist/lib/plugins/replication-graphql/index.js ./dist/lib/plugins/replication-websocket/index.js ./dist/lib/plugins/server-couchdb.js ./dist/lib/plugins/validate-z-schema.js ./dist/lib/plugins/lokijs/index.js ./dist/lib/plugins/dexie/index.js ./dist/lib/plugins/worker/index.js ./dist/lib/plugins/memory/index.js --no-dev --ignore-module util --ignore-module babel-plugin-transform-async-to-promises --ignore-module url --ignore-module ws --ignore-module graphql --ignore-module as-typed --ignore-module \"@types/*\"",
     "test:circular": "npm run build && madge --circular ./dist/es/index.js",
-    "test:performance:pouchdb": "npm run transpile && cross-env STORAGE=pouchdb mocha --config ./config/.mocharc.js ./test_tmp/performance.test.js --unhandled-rejections=strict --expose-gc",
-    "test:performance:lokijs": "npm run transpile && cross-env STORAGE=lokijs mocha --config ./config/.mocharc.js ./test_tmp/performance.test.js --unhandled-rejections=strict --expose-gc",
-    "test:performance:dexie-worker": "npm run transpile && cross-env STORAGE=dexie-worker mocha --config ./config/.mocharc.js ./test_tmp/performance.test.js --unhandled-rejections=strict --expose-gc",
+
+    "test:performance:pouchdb": "npm run transpile && cross-env DEFAULT_STORAGE=pouchdb CI=true karma start ./config/karma.performance.conf.js --single-run",
+    "test:performance:lokijs": "npm run transpile && cross-env DEFAULT_STORAGE=lokijs CI=true karma start ./config/karma.performance.conf.js --single-run",
+    "test:performance:dexie-worker": "npm run transpile && cross-env DEFAULT_STORAGE=dexie-worker CI=true karma start ./config/karma.performance.conf.js --single-run",
+    "test:performance:dexie": "npm run transpile && cross-env DEFAULT_STORAGE=dexie CI=true karma start ./config/karma.performance.conf.js --single-run",
+    "test:performance:memory": "npm run transpile && cross-env DEFAULT_STORAGE=memory mocha --config ./config/.mocharc.js ./test_tmp/performance.test.js --unhandled-rejections=strict --expose-gc",
+    "test:performance:foundationdb": "npm run transpile && cross-env DEFAULT_STORAGE=foundationdb mocha --config ./config/.mocharc.js ./test_tmp/performance.test.js --unhandled-rejections=strict --expose-gc",
+
+
+
     "couch:start": "docker run -d -p 5984:5984 --rm --name rxdb-couchdb couchdb:2.1.1",
     "couch:stop": "docker rm -f rxdb-couchdb",
     "test:couchdb": "npm run transpile && cross-env DEFAULT_STORAGE=pouchdb mocha --config ./config/.mocharc.js ./test_tmp/couch-db-integration.test.js",
